rename oneOf to fastOr
